Add selection-bar (better name, anyone?)
Fix up slider arrows so they may be recolored, and allowed coloring of
image-based widgets.

Removed the now-unnecessary optimisation of rendering stuff to texture.
image-based buttons with text are still rendered to texture, but that will
change when I implement the new text button code.

As a result of the refactor we can now have rotated text buttons ;)